tools/requirements.txt
httpx~=0.27.0
construct~=2.8.22
pyOpenSSL~=24.1.0
matplotlib~=3.9.0
numpy~=1.26
python-dotenv~=1.0.1
elasticsearch~=8.14.0
elasticsearch-dsl~=8.14.0